move init to util
