+++ rgf/rgf_model.py
@@ -17,7 +17,6 @@ from sklearn.utils.multiclass import check_classification_targets
 from sklearn.utils.validation import check_array, check_consistent_length, check_X_y, column_or_1d
 
 from rgf import util
-import rgf
 
 
 _ALGORITHMS = ("RGF", "RGF_Opt", "RGF_Sib")
@@ -475,15 +474,15 @@ class _RGFBinaryClassifier(BaseEstimator, ClassifierMixin):
         self.learning_rate = learning_rate
         self.memory_policy = memory_policy
         self.verbose = verbose
-        self._file_prefix = str(uuid4()) + str(rgf.COUNTER.increment())
-        rgf.UUIDS.append(self._file_prefix)
+        self._file_prefix = str(uuid4()) + str(util.COUNTER.increment())
+        util.UUIDS.append(self._file_prefix)
         self._fitted = None
         self._latest_model_loc = None
 
     def fit(self, X, y, sample_weight):
-        train_x_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".train.data.x")
-        train_y_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".train.data.y")
-        train_weight_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".train.data.weight")
+        train_x_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".train.data.x")
+        train_y_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".train.data.y")
+        train_weight_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".train.data.weight")
         if sp.isspmatrix(X):
             util.sparse_savetxt(train_x_loc, X)
         else:
@@ -517,10 +516,10 @@ class _RGFBinaryClassifier(BaseEstimator, ClassifierMixin):
         params.append("opt_interval=%s" % self.opt_interval)
         params.append("opt_stepsize=%s" % self.learning_rate)
         params.append("memory_policy=%s" % self.memory_policy.title())
-        params.append("model_fn_prefix=%s" % os.path.join(rgf.get_temp_path(), self._file_prefix + ".model"))
+        params.append("model_fn_prefix=%s" % os.path.join(util.get_temp_path(), self._file_prefix + ".model"))
         params.append("train_w_fn=%s" % train_weight_loc)
 
-        cmd = (rgf.get_exe_path(), "train", ",".join(params))
+        cmd = (util.get_exe_path(), "train", ",".join(params))
 
         # Train
         output = subprocess.Popen(cmd,
@@ -534,11 +533,11 @@ class _RGFBinaryClassifier(BaseEstimator, ClassifierMixin):
 
         self._fitted = True
         # Find latest model location
-        model_glob = os.path.join(rgf.get_temp_path(), self._file_prefix + ".model*")
+        model_glob = os.path.join(util.get_temp_path(), self._file_prefix + ".model*")
         model_files = glob(model_glob)
         if not model_files:
             raise Exception('Model learning result is not found in {0}. '
-                            'Training is abnormally finished.'.format(rgf.get_temp_path()))
+                            'Training is abnormally finished.'.format(util.get_temp_path()))
         self._latest_model_loc = sorted(model_files, reverse=True)[0]
         return self
 
@@ -547,22 +546,22 @@ class _RGFBinaryClassifier(BaseEstimator, ClassifierMixin):
             raise NotFittedError(util.not_fitted_error_desc())
         if not os.path.isfile(self._latest_model_loc):
             raise Exception('Model learning result is not found in {0}. '
-                            'This is rgf_python error.'.format(rgf.get_temp_path()))
+                            'This is rgf_python error.'.format(util.get_temp_path()))
 
-        test_x_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".test.data.x")
+        test_x_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".test.data.x")
         if sp.isspmatrix(X):
             util.sparse_savetxt(test_x_loc, X)
         else:
             np.savetxt(test_x_loc, X, delimiter=' ', fmt="%s")
 
         # Format test command
-        pred_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".predictions.txt")
+        pred_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".predictions.txt")
         params = []
         params.append("test_x_fn=%s" % test_x_loc)
         params.append("prediction_fn=%s" % pred_loc)
         params.append("model_fn=%s" % self._latest_model_loc)
 
-        cmd = (rgf.get_exe_path(), "predict", ",".join(params))
+        cmd = (util.get_exe_path(), "predict", ",".join(params))
 
         output = subprocess.Popen(cmd,
                                   stdout=subprocess.PIPE,
@@ -727,8 +726,8 @@ class RGFRegressor(util.RGFRegressorBase):
         self.learning_rate = learning_rate
         self.memory_policy = memory_policy
         self.verbose = verbose
-        self._file_prefix = str(uuid4()) + str(rgf.COUNTER.increment())
-        rgf.UUIDS.append(self._file_prefix)
+        self._file_prefix = str(uuid4()) + str(util.COUNTER.increment())
+        util.UUIDS.append(self._file_prefix)
         self._n_features = None
         self._fitted = None
         self._latest_model_loc = None
@@ -806,9 +805,9 @@ class RGFRegressor(util.RGFRegressorBase):
                 raise ValueError("Sample weights must be positive.")
         check_consistent_length(X, y, sample_weight)
 
-        train_x_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".train.data.x")
-        train_y_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".train.data.y")
-        train_weight_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".train.data.weight")
+        train_x_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".train.data.x")
+        train_y_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".train.data.y")
+        train_weight_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".train.data.weight")
         if sp.isspmatrix(X):
             util.sparse_savetxt(train_x_loc, X)
         else:
@@ -839,10 +838,10 @@ class RGFRegressor(util.RGFRegressorBase):
         params.append("opt_interval=%s" % self.opt_interval)
         params.append("opt_stepsize=%s" % self.learning_rate)
         params.append("memory_policy=%s" % self.memory_policy.title())
-        params.append("model_fn_prefix=%s" % os.path.join(rgf.get_temp_path(), self._file_prefix + ".model"))
+        params.append("model_fn_prefix=%s" % os.path.join(util.get_temp_path(), self._file_prefix + ".model"))
         params.append("train_w_fn=%s" % train_weight_loc)
 
-        cmd = (rgf.get_exe_path(), "train", ",".join(params))
+        cmd = (util.get_exe_path(), "train", ",".join(params))
 
         # Train
         output = subprocess.Popen(cmd,
@@ -857,11 +856,11 @@ class RGFRegressor(util.RGFRegressorBase):
         self._fitted = True
 
         # Find latest model location
-        model_glob = os.path.join(rgf.get_temp_path(), self._file_prefix + ".model*")
+        model_glob = os.path.join(util.get_temp_path(), self._file_prefix + ".model*")
         model_files = glob(model_glob)
         if not model_files:
             raise Exception('Model learning result is not found in {0}. '
-                            'Training is abnormally finished.'.format(rgf.get_temp_path()))
+                            'Training is abnormally finished.'.format(util.get_temp_path()))
         self._latest_model_loc = sorted(model_files, reverse=True)[0]
         return self
 
@@ -892,7 +891,7 @@ class RGFRegressor(util.RGFRegressorBase):
                              "input n_features is %s "
                              % (self._n_features, n_features))
 
-        test_x_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".test.data.x")
+        test_x_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".test.data.x")
         if sp.isspmatrix(X):
             util.sparse_savetxt(test_x_loc, X)
         else:
@@ -900,16 +899,16 @@ class RGFRegressor(util.RGFRegressorBase):
 
         if not os.path.isfile(self._latest_model_loc):
             raise Exception('Model learning result is not found in {0}. '
-                            'This is rgf_python error.'.format(rgf.get_temp_path()))
+                            'This is rgf_python error.'.format(util.get_temp_path()))
 
         # Format test command
-        pred_loc = os.path.join(rgf.get_temp_path(), self._file_prefix + ".predictions.txt")
+        pred_loc = os.path.join(util.get_temp_path(), self._file_prefix + ".predictions.txt")
         params = []
         params.append("test_x_fn=%s" % test_x_loc)
         params.append("prediction_fn=%s" % pred_loc)
         params.append("model_fn=%s" % self._latest_model_loc)
 
-        cmd = (rgf.get_exe_path(), "predict", ",".join(params))
+        cmd = (util.get_exe_path(), "predict", ",".join(params))
 
         output = subprocess.Popen(cmd,
                                   stdout=subprocess.PIPE,
